backend/app/klipperscmd/pyproject.toml
[project]
name = "klipperscmd"
version = "0.1.0"
description = ""
authors = [{name = "halilagin", email = "halil.agin@gmail.com"}]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "pydantic[email]>=2.10.6",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.6",
    "python-dotenv==1.0.0",
    "pydantic-settings==2.1.0",
    "httpx>=0.28.1",
    "pillow>=11.2.1",
    "pytz>=2025.2",
    "openai>=2.8.0",
    "pydub>=0.25.1",
    "grpcio==1.62.2",
    "aiofiles>=24.1.0",
    "psutil>=7.0.0",
    "requests>=2.31.0",
    "ffmpeg-python>=0.2.0",
    "pysubs2>=1.8.0",
    "boto3>=1.40.45",
    "click>=8.3.0",
    "redis>=7.1.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["clippercmd*", "clippertool*", "db*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
